Move PopUpMessagesContainer and related components to new structure for improved organization.
